Changes - chage type column in records_services

+++ database/migrations/2019_08_19_000000_create_records_services_table.php
@@ -20,7 +20,7 @@ class CreateRecordsServicesTable extends Migration
             $table->string('name')->nullable();
             $table->string('description')->nullable();
             $table->integer('price')->nullable();
-            $table->integer('amount')->nullable();
+            $table->float('amount', 8, 2)->nullable();
             $table->integer('total')->nullable();
             $table->timestamps();
         });
